Completed Functional Freebase Import Job
Job now fully functional, with batching, bulk insert, and creation of
proper relational data in Books, Authors, Genres, and BookGenres tables.
May need tweaking as we go, and my merges aren't the most elegant (using
distinct and some split strings...) but it works and I'm not so big on
the SQL side tonight, so yay.